Minecraft Map: Make note of where you've been
Minecraft Map will allow you to make a map. Specify coordinates in the
upper boxes or mark directly to the map. You can choose from a variety
of stamp-like images to stick onto the map. If you found a desert temple
at coordinates (1000,1000), you can mark a picture of a desert temple on
that coordinate location on the map.